a lot of cameraKit stuff

+++ src/CameraKit/Sources/CameraKit/Datatypes/CKAdjustableCameraConfiguration.swift
@@ -5,7 +5,7 @@ public struct CKAdjustableCameraConfiguration: Identifiable, Hashable {
   public let maxZoom: Double
   public let minFps: Int
   public let maxFps: Int
-  public let fieldOfView: Float
+  public let fieldOfView: Double
   public let isVideoMirroringAvailable: Bool
   public let supportedStabilizationModes: [CKStabilizationMode]
   public let isMulticamAvailable: Bool
